+++ src/Http/Middleware/SetLocale.php
@@ -4,6 +4,7 @@ namespace BamboleeDigital\TranslatableResourceKit\Http\Middleware;
 
 use Closure;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Log;
 use Symfony\Component\HttpFoundation\Response;
 
 /**
@@ -25,12 +26,19 @@ class SetLocale
      */
     public function handle(Request $request, Closure $next): Response
     {
+        if (config('translatable-resource-kit.debug', false)) {
+            Log::info('Setting locale');
+        }
         // Get the requested language
         $locale = $request->input('lang', config('app.locale'));
 
         // Check if the requested language is supported, otherwise set the fallback
         if (!in_array($locale, config('translatable-resource-kit.supported_locales', [config('app.fallback_locale', 'en')]))) {
             $locale = config('app.fallback_locale', 'en');
+        }
+        
+        if (config('translatable-resource-kit.debug', false)) {
+            Log::info('Locale set to: ', ['locale' => $locale]);
         }
 
         // Set the application locale